+++ src/apiImplementation.ts
@@ -1,10 +1,5 @@
-
-
 import { RHelp } from './rHelp';
-
 import { RExtension } from './api';
-
-
 export class RExtensionImplementation implements RExtension  {
     public helpPanel?: RHelp;
 }